refactor(front): extract flow canvas into FlowCanvas component and consolidate canvas logic
